+++ src/MythicScribe.ts
@@ -10,16 +10,18 @@ import {
 import { ScribeMechanicHandler } from './datasets/ScribeMechanic';
 import { ScribeEnumHandler } from './datasets/ScribeEnum';
 import { doVersionSpecificMigrations } from './migration/migration';
-import { showInfoMessageWithOptions } from './utils/logger';
+import { logDebug, logsProvider, openLogs, showInfoMessageWithOptions } from './utils/logger';
 
 export let ctx: vscode.ExtensionContext;
 
 export async function activate(context: vscode.ExtensionContext) {
     ctx = context;
+    logDebug('Extension Activated');
 
     // Check if the extension has been updated
     if (checkExtensionVersion()) {
         // Run migrations if so
+        logDebug('Running migrations');
         await doVersionSpecificMigrations();
     }
 
@@ -36,9 +38,13 @@ export async function activate(context: vscode.ExtensionContext) {
         vscode.commands.registerCommand('MythicScribe.addCustomDataset', addCustomDataset),
         vscode.commands.registerCommand('MythicScribe.removeCustomDataset', removeCustomDataset),
         vscode.commands.registerCommand('MythicScribe.createBundleDataset', createBundleDataset),
+        vscode.commands.registerCommand('MythicScribe.openLogs', openLogs),
 
         // Formatter
-        getFormatter()
+        getFormatter(),
+
+        // Logger
+        vscode.workspace.registerTextDocumentContentProvider('mythicscribelogs', logsProvider)
     );
 
     if (vscode.window.activeTextEditor) {
@@ -51,6 +57,7 @@ export function deactivate() {}
 export function checkExtensionVersion(): boolean {
     const version = ctx.extension.packageJSON.version;
     const savedVersion = ctx.globalState.get<string>('extensionVersion');
+    logDebug(`Current version: ${version}, Saved version: ${savedVersion}`);
     if (version && version !== savedVersion) {
         const checkExtensionVersionOptions: { [key: string]: string } = {
             'Check Changelogs': 'https://github.com/Lxlp38/MythicScribe/blob/master/CHANGELOG.md',