+++ src/subscriptions/SubscriptionHandler.ts
@@ -28,6 +28,7 @@ import { ctx } from '../MythicScribe';
 import { genericFileCompletionProvider } from '../completions/filecompletions/genericFileCompletionProvider';
 import { DroptableFileObject } from '../schemas/droptableFileObjects';
 import { StatFileObjects } from '../schemas/statfileObjects';
+import { logDebug } from '../utils/logger';
 
 type SubscriptionFunction = () => vscode.Disposable;
 type SubscriptionCondition = () => boolean;
@@ -69,7 +70,7 @@ export abstract class AbstractScribeSubscription {
      */
     enableAll() {
         this.disposeAll();
-
+        logDebug('Enabling all subscriptions for', this.constructor.name);
         this.childSubscriptionHandlers.forEach((handler) => {
             handler.enableAll();
         });
@@ -94,6 +95,10 @@ export abstract class AbstractScribeSubscription {
             handler.disposeAll();
         });
 
+        if (this.subscriptions.length === 0) {
+            return;
+        }
+        logDebug('Disposing all subscriptions for', this.constructor.name);
         this.subscriptions.forEach((subscription) => {
             subscription.dispose();
         });