+++ src/utils/logger.ts
@@ -1,5 +1,68 @@
 import * as vscode from 'vscode';
 
+export const logs: string[] = [];
+
+export enum LogType {
+    INFO = 'INFO',
+    WARNING = 'WARNING',
+    ERROR = 'ERROR',
+    DEBUG = 'DEBUG',
+}
+
+const virtualLogsUri = vscode.Uri.parse('mythicscribelogs:/Logs.log');
+let updateTimeout: NodeJS.Timeout;
+export let isLogFileOpen = false;
+
+/**
+ * Logs a message with metadata including a timestamp and log type.
+ *
+ * @param message - The message to log.
+ * @param type - The type of log message. Defaults to `LogType.DEBUG`.
+ */
+export function logMetadata(message: string, type: LogType = LogType.DEBUG) {
+    const timestamp = new Date().toISOString();
+    const formattedMessage = `[${timestamp}] ${type}: ${message}`;
+    logs.push(formattedMessage);
+    if (!isLogFileOpen) {
+        return;
+    }
+    clearTimeout(updateTimeout);
+    updateTimeout = setTimeout(() => {
+        logsProvider.update(vscode.Uri.parse('mythicscribelogs:/Logs.log'));
+    }, 500);
+}
+
+/**
+ * Opens the logs by updating the logs provider and displaying the logs in a text document.
+ *
+ * This function updates the logs provider with the virtual logs URI and then opens the logs
+ * in a new text document within the VSCode editor. The document is opened in non-preview mode.
+ * Once the document is opened, it sets the `isLogFileOpen` flag to true.
+ *
+ * @returns {Promise<void>} A promise that resolves when the logs are opened.
+ */
+export async function openLogs(): Promise<void> {
+    logsProvider.update(virtualLogsUri);
+    vscode.window.showTextDocument(virtualLogsUri, { preview: false }).then(() => {
+        isLogFileOpen = true;
+        return;
+    });
+}
+
+class LogsProvider implements vscode.TextDocumentContentProvider {
+    private _onDidChange = new vscode.EventEmitter<vscode.Uri>();
+    onDidChange?: vscode.Event<vscode.Uri> = this._onDidChange.event;
+
+    provideTextDocumentContent(): vscode.ProviderResult<string> {
+        return logs.join('\n');
+    }
+
+    public update(uri: vscode.Uri) {
+        this._onDidChange.fire(uri);
+    }
+}
+export const logsProvider = new LogsProvider();
+
 /**
  * Logs an error message to the Visual Studio Code error message window.
  *
@@ -7,6 +70,7 @@ import * as vscode from 'vscode';
  * @param message - An optional custom message to display before the error message. Defaults to 'An error occurred'.
  */
 export function logError(error: unknown, message: string = 'An error occurred:') {
+    logMetadata(message, LogType.ERROR);
     if (error instanceof Error) {
         vscode.window.showErrorMessage(message + '\n' + error);
     } else {
@@ -20,6 +84,7 @@ export function logError(error: unknown, message: string = 'An error occurred:')
  * @param message - The warning message to be displayed.
  */
 export function logWarning(message: string) {
+    logMetadata(message, LogType.WARNING);
     vscode.window.showWarningMessage(message);
 }
 
@@ -29,7 +94,12 @@ export function logWarning(message: string) {
  * @param message - The message to be displayed.
  */
 export function logInfo(message: string) {
+    logMetadata(message, LogType.INFO);
     vscode.window.showInformationMessage(message);
+}
+
+export function logDebug(...message: string[]) {
+    logMetadata(message.join(' '), LogType.DEBUG);
 }
 
 /**
@@ -45,9 +115,11 @@ export async function showInfoMessageWithOptions(
     message: string,
     options: { [key: string]: string }
 ) {
+    logMetadata(message, LogType.DEBUG);
     const optionKeys = Object.keys(options);
     return vscode.window.showInformationMessage(message, ...optionKeys).then((selected) => {
         if (selected) {
+            logMetadata(`Opened ${options[selected]}`, LogType.DEBUG);
             return vscode.env.openExternal(vscode.Uri.parse(options[selected]));
         }
         return undefined;